bytes: move crc utilities to hashing library
Keeping the iobuf-bytes crc utilities in the bytes library cause awkward
circular-ish dependencies. In reality bytes and hashing are fundamental
enough they should probably live in a library common to the entire tree.
As long as we are going to keep such fine-grained libraries, then they
should also stay lean and specific.
