+++ src/v/bytes/utils.h
@@ -11,16 +11,6 @@
 
 #pragma once
 #include "bytes/iobuf.h"
-#include "hashing/crc32c.h"
-
-inline void crc_extend_iobuf(crc::crc32c& crc, const iobuf& buf) {
-    auto in = iobuf::iterator_consumer(buf.cbegin(), buf.cend());
-    (void)in.consume(buf.size_bytes(), [&crc](const char* src, size_t sz) {
-        // NOLINTNEXTLINE
-        crc.extend(reinterpret_cast<const uint8_t*>(src), sz);
-        return ss::stop_iteration::no;
-    });
-}
 
 inline bool is_zero(const char* data, size_t size) {
     // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
