+++ src/v/hashing/crc32c.h
@@ -10,6 +10,8 @@
  */
 
 #pragma once
+#include "bytes/iobuf.h"
+
 #include <crc32c/crc32c.h>
 
 #include <type_traits>
@@ -40,3 +42,12 @@ private:
 };
 
 } // namespace crc
+
+inline void crc_extend_iobuf(crc::crc32c& crc, const iobuf& buf) {
+    auto in = iobuf::iterator_consumer(buf.cbegin(), buf.cend());
+    (void)in.consume(buf.size_bytes(), [&crc](const char* src, size_t sz) {
+        // NOLINTNEXTLINE
+        crc.extend(reinterpret_cast<const uint8_t*>(src), sz);
+        return ss::stop_iteration::no;
+    });
+}
