+++ src/v/storage/snapshot.cc
@@ -10,7 +10,6 @@
 #include "storage/snapshot.h"
 
 #include "bytes/iobuf_parser.h"
-#include "bytes/utils.h"
 #include "hashing/crc32c.h"
 #include "random/generators.h"
 #include "reflection/adl.h"
